(BOLT-1103) Validate action parameter for bash/powershell
The enable and disable actions are not allowed in bash/powershell because the behavior with different init systems is too complicated/unreliable. In the case where the `service::linux` or `service::windows` task is invoked directly the metadata validation will handle this, however if the task is invoked with just `service` and the `puppet-agent` feature is not set on the target and the `shell` or `powershell` feature is then those actions will not be forbidden by metatdata. This commit implements verification that only the available actions are allowed.