(maint) Update service task return to use "enable"
Since `puppet resource service <service_name>` settings are called
`ensure` and `enable`, i.e.
service { 'puppet':
  ensure => 'running',
  enable => 'true',
}
it would be nice to have the service task return similarly named
parameters. This would allow using the result of a `service` task call
within a Bolt plan to be used in an `apply` block. As it stands now, the
result of a `service` call with break in an apply block because
`enabled` is not a valid parameter, but `enable` is.