(MAINT) Fix debian 8 acceptance tests
Prior to this commit the acceptance tests for Debian 8 would
fail on calling stop for the rsyslog service due to a weird
inconsistency that causes an error if systemctl stop is called
after systemctl start if the service has been enabled beforehand.

These acceptance tests were not running against Debian 8 prior
to the litmus conversion which is potentially why we did not
catch this.

This commit updates the tests to use apache2 instead of rsyslog
and includes a minor update to the regex to accommodate ubuntu
14.04 which is EOL, but still supported. When that support is
removed the regex can be simplified.