(MAINT) Remove puppet 5 from spec matrix
This commit removes puppet 5 from the parallel spec matrix as
there is a dependency conflict as litmus requires bolt which
then requires puppet 6 or greater. The acceptance tests do
run against nodes with puppet 5 agents installed on them and,
in this case, there are no unit tests running so none are
actually skipped.