(FM-8342) Enable mocked tests on localhost
This commit updates the spec_helper_acceptance to use the new
helpers and features in litmus to be take advantage of the
ability to mock features for localhost.

It also updates the test files to remove branching behavior
for testing between localhost and remote hosts, as well as
removing the now unneccessary skips.