(MODULES-9979) Fix empty return values in Linux task
Prior to this commit, querying for a nonexistent service using some
service managers in the Linux task would return empty values.  This
commit also captures stderr in status checks to accurately convey that
the service is unknown (or any other error).

This also breaks out service and initctl into their own case statements,
as well as fixes a couple of minor bugs around trying to start or stop
a service already in the desired state returning an error.