(maint) Fix up accentance testing
Updating facter affected which family Ubuntu 14.04 was reporting, which
impacted whether the platform was skipped for service 'enable' task
testing. Separately there was a new, different but still successful
message for Ubuntu 16.04 platform for the service enable tests. This
updates tests to correctly skip Ubuntu 14.04 tests and updates the
expected message on Ubuntu 16.04.